[REF] web_editor, website, *: move assets method to website
*: website_sale

This commit prepares web_editor deletion.
