+++ addons/website/models/theme_models.py
@@ -254,7 +254,7 @@ class ThemeUtils(models.AbstractModel):
     @api.model
     def _reset_default_config(self):
         # Reinitialize some css customizations
-        self.env['web_editor.assets'].make_scss_customization(
+        self.env['website.assets'].make_scss_customization(
             '/website/static/src/scss/options/user_values.scss',
             {
                 'font': 'null',
