+++ addons/website/static/src/components/resource_editor/resource_editor.js
@@ -297,7 +297,7 @@ export class ResourceEditor extends Component {
             throw new Error(_t("Reseting views is not supported yet"));
         }
         const resource = this.state.currentResource;
-        await this.orm.call("web_editor.assets", "reset_asset", [resource.url, resource.bundle], {
+        await this.orm.call("website.assets", "reset_asset", [resource.url, resource.bundle], {
             context: this.context,
         });
         await this.loadResources();
@@ -367,7 +367,7 @@ export class ResourceEditor extends Component {
             : this.state.resources.scss[url].bundle;
         const fileType = isJSFile ? "js" : "scss";
         const params = [url, bundle, arch, fileType];
-        await this.orm.call("web_editor.assets", "save_asset", params, { context: this.context });
+        await this.orm.call("website.assets", "save_asset", params, { context: this.context });
         delete resource.dirty;
     }
